Refactors 1- Renamed package com/mrf/javadecompiler/openapi/ to com/mrf/javadecompiler/openide/
2- Deprecate com.mrf.javadecompiler.builder.SourceWindowBuilder

3- Fix comment on class
com.mrf.javadecompiler.openapi.ui.SourceWindowTopComponent
