+++ src/main/java/com/mrf/javadecompiler/openide/action/DecompileAction.java
@@ -14,7 +14,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.mrf.javadecompiler.openapi.action;
+package com.mrf.javadecompiler.openide.action;
 
 import static com.machinezoo.noexception.Exceptions.wrap;
 import static com.mrf.javadecompiler.constants.Constants.HEADER_COMMENT;
