+++ src/main/nbm/manifest.mf
@@ -1,5 +1,5 @@
 Manifest-Version: 1.0
 OpenIDE-Module-Localizing-Bundle: com/mrf/javadecompiler/Bundle.properties
 OpenIDE-Module-Requires: org.openide.windows.WindowManager
-OpenIDE-Module-Install: com/mrf/javadecompiler/openapi/Installer.class
+OpenIDE-Module-Install: com/mrf/javadecompiler/openide/Installer.class
